add pdf, editText
Добавил пдф и редактирование и сохранение

+++ InfoFolder/StretchCeilingCalculator/app/src/main/java/bokarev/st/stretchceilingcalculator/entities/Estimate.kt
@@ -10,7 +10,7 @@ data class Estimate(
     val _id: Int,
     val _idClient: Int,
     val _idTypeCategory: Int,
-    val Count: Int,
+    val Count: Double,
     val DateOfCreation: String,
     val DateOfEditing: String,
 )
